feat: Rework the storage model
Key Changes:
Switch from using a single btreemap for tasks storage to using an array
of tasks and a one to one mapping of tasks id's and the task's index
slot in the array.

This change allows for quick reindexing on delete by eliminating the
need to clone Tasks stored in the btreemap as the previous
implementation.